github_latest_version/requirements.txt
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
Deprecated==1.2.13
greenlet==2.0.2
idna==3.4
pycparser==2.21
PyGithub==1.58.0
PyJWT==2.6.0
PyNaCl==1.5.0
requests==2.28.2
SQLAlchemy==2.0.4
typing_extensions==4.5.0
urllib3==1.26.14
wrapt==1.14.1
pyverilog==1.3.0
argparse==1.4.0
